back-end/requirements.txt
alembic==1.4.0
antiorm==1.2.1
attrs==19.3.0
automium==0.2.6
automium-web==0.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
click-option-group==0.3.0
config==0.4.2
db==0.1.1
decorator==4.4.1
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
legacy==0.1.6
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-sqlalchemy==0.22.2
mock==4.0.1
more-itertools==8.2.0
MyApplication==0.1.0
numpy==1.18.1
packaging==20.1
pandas==1.0.1
packaging==20.1
pathfinding==0.0.4
pluggy==0.13.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
public==2019.4.13
py==1.8.1
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyswip==0.2.9
pytest==5.3.5
pytest-base-url==1.4.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
self==2019.4.13
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.7
wcwidth==0.1.8
Werkzeug==1.0.0
zipp==3.0.0
